refactor(CLI & Config): Flatten Rust file tree
